+++ docs/changelog.md
@@ -4,6 +4,7 @@
 
 **Features**:
 
++ "Relocate" dialog options for editing point position.
 + Fix commas error of expression table on GUI.
 + Fix "Add Point" action on menu bar.